PICARD-2213: Move classes for script serialization into submodule
